third_party/blink/web_tests/storage/domstorage/localstorage/missing-arguments-expected.txt

Ensures that we get exceptions when we do not include required arguments for localStorage methods.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS localStorage.key() threw exception TypeError: Failed to execute 'key' on 'Storage': 1 argument required, but only 0 present..
PASS localStorage.getItem() threw exception TypeError: Failed to execute 'getItem' on 'Storage': 1 argument required, but only 0 present..
PASS localStorage.setItem() threw exception TypeError: Failed to execute 'setItem' on 'Storage': 2 arguments required, but only 0 present..
FAIL localStorage.setItem("a") should throw TypeError: Failed to execute 'setItem' on 'Storage': 2 arguments required, but only 0 present.. Threw exception TypeError: Failed to execute 'setItem' on 'Storage': 2 arguments required, but only 1 present..
PASS localStorage.removeItem() threw exception TypeError: Failed to execute 'removeItem' on 'Storage': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
